# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RoadTrafficManagement, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Personnel/GetById/{Id}": {
      "get": {
        "tags": [
          "Personnel"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Personnel/GetAll/{startIndex}/{count}": {
      "get": {
        "tags": [
          "Personnel"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "startIndex",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "count",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Personnel/Search": {
      "get": {
        "tags": [
          "Personnel"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PersonnelCriteria"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Personnel/Create": {
      "post": {
        "tags": [
          "Personnel"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PersonnelRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Personnel/Update": {
      "post": {
        "tags": [
          "Personnel"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PersonnelRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Personnel/Delete": {
      "post": {
        "tags": [
          "Personnel"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PersonnelRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Road/GetById/{Id}": {
      "get": {
        "tags": [
          "Road"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Road/GetAll/{startIndex}/{count}": {
      "get": {
        "tags": [
          "Road"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "startIndex",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "count",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Road/Search": {
      "get": {
        "tags": [
          "Road"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RoadCriteria"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Road/Create": {
      "post": {
        "tags": [
          "Road"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RoadRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Road/Update": {
      "post": {
        "tags": [
          "Road"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RoadRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Road/Delete": {
      "post": {
        "tags": [
          "Road"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RoadRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RoadProperty/GetById/{Id}": {
      "get": {
        "tags": [
          "RoadProperty"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "Id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RoadProperty/GetAll/{startIndex}/{count}": {
      "get": {
        "tags": [
          "RoadProperty"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "startIndex",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "count",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RoadProperty/Search": {
      "post": {
        "tags": [
          "RoadProperty"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RoadPropertyCriteria"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RoadProperty/Create": {
      "post": {
        "tags": [
          "RoadProperty"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RoadPropertyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RoadProperty/Update": {
      "post": {
        "tags": [
          "RoadProperty"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RoadPropertyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RoadProperty/Delete": {
      "post": {
        "tags": [
          "RoadProperty"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RoadPropertyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "LoginModel": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "PersonnelCriteria": {
      "type": "object",
      "properties": {
        "startIndex": {
          "format": "int32",
          "type": "integer"
        },
        "count": {
          "format": "int32",
          "type": "integer"
        },
        "personelCode": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "mobilePhoneNumber": {
          "type": "string"
        },
        "department": {
          "type": "string"
        },
        "employmentType": {
          "type": "string"
        }
      }
    },
    "PersonnelRequest": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "personelCode": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "mobilePhoneNumber": {
          "type": "string"
        },
        "department": {
          "type": "string"
        },
        "employmentType": {
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "modifiedDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "RegisterModel": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
...
    "RoadCriteria": {
      "type": "object",
      "properties": {
        "startIndex": {
          "format": "int32",
          "type": "integer"
        },
        "count": {
          "format": "int32",
          "type": "integer"
        },
        "typeId": {
          "format": "int32",
          "type": "integer"
        },
        "parentId": {
          "format": "int32",
          "type": "integer"
        },
        "roadCode": {
          "type": "string"
        },
        "roadName": {
          "type": "string"
        },
        "chainageFrom": {
          "type": "string"
        },
        "chainageTo": {
          "type": "string"
        },
        "kilometerStart": {
          "format": "double",
          "type": "number"
        },
        "kilometerEnd": {
          "format": "double",
          "type": "number"
        },
        "startGps": {
          "type": "string"
        },
        "endGps": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
...
    "RoadPropertyCriteria": {
      "type": "object",
      "properties": {
        "startIndex": {
          "format": "int32",
          "type": "integer"
        },
        "count": {
          "format": "int32",
          "type": "integer"
        },
        "propertyCode": {
          "type": "string"
        },
        "propertyName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "modifiedDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "RoadPropertyRequest": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "propertyCode": {
          "type": "string"
        },
        "propertyName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "modifiedDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "RoadRequest": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "typeId": {
          "format": "int32",
          "type": "integer"
        },
        "parentId": {
          "format": "int32",
          "type": "integer"
        },
        "roadCode": {
          "type": "string"
        },
        "roadName": {
          "type": "string"
        },
        "chainageFrom": {
          "type": "string"
        },
        "chainageTo": {
          "type": "string"
        },
        "startGps": {
          "type": "string"
        },
        "kilometerStart": {
          "format": "double",
          "type": "number"
        },
        "kilometerEnd": {
          "format": "double",
          "type": "number"
        },
        "endGps": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        },
        "modifiedDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
  }
}